scripts/requirements.txt
confidence
psycopg2-binary
pytz

